Improve Exception handling and Documentation readme
