+++ src/PublicSuffixList.php
@@ -16,7 +16,7 @@ interface PublicSuffixList extends DomainResolver, JsonSerializable
     /**
      * Returns PSL info for a given domain against the PSL rules for cookie domain detection.
      *
-     * @throws CannotResolveDomain   if the domain is invalid
+     * @throws SyntaxError           if the domain is invalid
      * @throws UnableToResolveDomain if the effective TLD can not be resolved
      */
     public function getCookieDomain(Host $host): ResolvedDomainName;
@@ -24,7 +24,7 @@ interface PublicSuffixList extends DomainResolver, JsonSerializable
     /**
      * Returns PSL info for a given domain against the PSL rules for ICANN domain detection.
      *
-     * @throws CannotResolveDomain   if the domain is invalid
+     * @throws SyntaxError           if the domain is invalid
      * @throws UnableToResolveDomain if the domain does not contain a ICANN Effective TLD
      */
     public function getICANNDomain(Host $host): ResolvedDomainName;
@@ -32,7 +32,7 @@ interface PublicSuffixList extends DomainResolver, JsonSerializable
     /**
      * Returns PSL info for a given domain against the PSL rules for private domain detection.
      *
-     * @throws CannotResolveDomain   if the domain is invalid
+     * @throws SyntaxError           if the domain is invalid
      * @throws UnableToResolveDomain if the domain does not contain a private Effective TLD
      */
     public function getPrivateDomain(Host $host): ResolvedDomainName;
