+++ src/RulesTest.php
@@ -354,7 +354,7 @@ final class RulesTest extends TestCase
      */
     public function testGetPublicSuffixThrowsInvalidDomainException(string $domain): void
     {
-        self::expectException(CannotResolveDomain::class);
+        self::expectException(SyntaxError::class);
 
         $this->rules->getICANNDomain($domain);
     }
@@ -376,7 +376,7 @@ final class RulesTest extends TestCase
      */
     public function testGetPublicSuffixThrowsInvalidHostException(string $domain): void
     {
-        self::expectException(CannotResolveHost::class);
+        self::expectException(SyntaxError::class);
 
         $this->rules->getICANNDomain($domain);
     }
